separated the files for routes
