+++ routes/users.js
@@ -5,7 +5,7 @@ const bcrypt = require("bcryptjs");
 const jwt = require("jsonwebtoken");
 const { v4 } = require("uuid");
 
-const { Users, Courses } = require("./../database/db");
+const { Users } = require("./../database/db");
 const { ObjectID } = require("mongodb");
 
 router.get("/test", (req, res) => {
@@ -98,198 +98,7 @@ const verifyUser = async (req, res, next) => {
   }
 
 };
-router.get("/testOP", verifyUser, (req, res, next) => {
-  console.log(req.body);
-  res.status(200).json({ mo: "lo" });
-});
 
-//Course
-router.post("/getCourses", verifyUser, async (req, res, next) => {
-  if (!req.body.verified) {
-    next(err);
-    throw new Error("not verified");
-  }
-  try {
-    const courses = await Courses.find({ userId: req.body._id }).toArray();
-    if (courses) {
-	     return res.status(200).json(courses);
-    }
-    else {
-      return res.status(404).json({"msg": "no courses found"})
-    }
-  } catch (err) {
-    next(err);
-    console.log(err.stack);
-    throw new Error("Database is having a problem");
-  }
-});
-router.get("/getCoursesbyName", verifyUser, async (req, res, next) => {
-  if (!req.body.verified) {
-    next(err);
-    throw new Error("not verified");
-  }
-  try {
-    const courses = await Courses.find({userId: req.body._id}).toArray();
-    if (courses) {
-
-      return res.status(200).json(courses);
-    }
-    else {
-      return res.status(404).json({"msg": "can't find course"});
-    }
-  }
-  catch (err) {
-    next(err);
-    console.log(err.stack);
-    throw new Error("Database is having a problem");
-  }
-});
-
-
-router.post("/addCourse", verifyUser, async (req, res, next) => {
-  if (!req.body.verified) {
-    next(err);
-    throw new Error("not verified");
-  }
-  let course = {
-    courseName: req.body.courseName,
-    courseLink: req.body.courseLink,
-    description: req.body.description,
-    rating: req.body.rating,
-    resourcesLink: req.body.resources,
-    notes: [],
-    userId: req.body._id,
-  };
-  try {
-    const result = await Courses.insertOne(course);
-  } catch (error) {
-    next(err);
-    throw new Error("Database is having a problem");
-  }
-  res.status(201).json({ msg: "course added succesfully" });
-});
-
-router.post("/deleteCourses", verifyUser, async (req, res, next) => {
-  if (!req.body.verified) {
-    next(err);
-    throw new Error("not verified");
-  }
-
-  try {
-    const result = await Courses.deleteOne({
-      _id: new ObjectID(req.body.courseId),
-    });
-  } catch (error) {
-    next(err);
-    throw new Error("Database is having a problem");
-  }
-  res.status(201).json({ msg: "course added succesfully" });
-});
-
-router.post("/updateCourses", verifyUser, async (req, res, next) => {
-  if (!req.body.verified) {
-    next(err);
-    throw new Error("not verified");
-  }
-
-  try {
-    const result = await Courses.updateOne(
-      { _id: new ObjectID(req.body.courseId) },
-      {
-        $set: {
-          courseName: req.body.courseName,
-          courseLink: req.body.courseLink,
-          resourcesLink: req.body.resources,
-          description: req.body.description,
-          rating: req.body.rating,
-        },
-      }
-    );
-  } catch (error) {
-    next(err);
-    throw new Error("Database is having a problem");
-  }
-  res.status(201).json({ msg: "course added succesfully" });
-});
-
-//Notes
-
-router.get("/getNotes", verifyUser, async (req, res, next) => {
-  if (!req.body.verified) {
-    next(err);
-    throw new Error("not verified");
-  }
-  try {
-    let enotes = await Courses.findOne({
-      _id: new ObjectID(req.body.courseId),
-    });
-    const notes = enotes.notes;
-    res.status(200).json({ msg: "found", notes });
-  } catch (err) {
-    next(err);
-    throw new Error("Database is having a problem");
-  }
-});
-
-router.post("/addNotes", verifyUser, async (req, res, next) => {
-  if (!req.body.verified) {
-    next(err);
-    throw new Error("not verified");
-  }
-  let courseId = req.body.courseId;
-  let note = {
-    noteId: v4(),
-    text: req.body.text,
-    timeStamp: req.body.timeStamp,
-    heading: req.body.heading,
-  };
-  try {
-    const result = await Courses.updateOne(
-      { _id: new ObjectID(courseId) },
-      { $push: { notes: note } }
-    );
-  } catch (err) {
-    next(err);
-    throw new Error("Database is having a problem");
-  }
-});
-router.post("/deleteNotes", verifyUser, async (req, res, next) => {
-  try {
-    if (!req.body.verified) {
-      next(err);
-      throw new Error("not verified");
-    }
-    const result = Courses.updateOne(
-      { _id: new ObjectID(req.body.courseId) },
-      { $pull: { notes: { $elematch: { noteId: req.body.noteId } } } }
-    );
-    res.status(200).json({ msg: "done" });
-  } catch (err) {
-    next(err);
-    throw new Error("not verified");
-  }
-});
-router.post("/updateNotes", verifyUser, (req, res, next) => {
-  try {
-    if (!req.body.verified) {
-      next(err);
-      throw new Error("not verified");
-    }
-    const result = Courses.updateOne(
-      { _id: new ObjectID(req.body.courseId), "notes.noteId": req.body.noteId },
-      {
-        $set: {
-          "notes.$.heading": req.body.heading,
-          "notes.$.text": req.body.text,
-        },
-      }
-    );
-    res.status(200).json({ msg: "done" });
-  } catch (err) {
-    next(err);
-    throw new Error("not verified");
-  }
-});
 
 router.use(function (err, req, res, next) {
     res.status(500).send({ err: "Some error occured" });
